♻️ Refactor: apis.ts 파일 통합
