--- src/features/my-page/my-activity/apis.ts
@@ -1,20 +1,0 @@
-import {
-  GetMyActivitiesQuerySchema,
-  GetMyActivitiesResponseSchema,
-  type GetMyActivitiesQuery,
-  type GetMyActivitiesResponse,
-} from '@/features/my-page/types';
-import { bffFetch } from '@/shared/apis/base/bffFetch';
-
-const EMPTY_MY_ACTIVITIES_RESPONSE: GetMyActivitiesResponse = {
-  cursorId: null,
-  totalCount: 0,
-  activities: [],
-};
-
-export const getMyActivities = async (query: GetMyActivitiesQuery = {}) => {
-  const safeQuery = GetMyActivitiesQuerySchema.parse(query);
-  const result = await bffFetch.get<GetMyActivitiesResponse>('/my-activities', safeQuery);
-  if (!result) return EMPTY_MY_ACTIVITIES_RESPONSE;
-  return GetMyActivitiesResponseSchema.parse(result);
-};
