+++ src/features/my-page/my-activity/queries/useMyActivities.ts
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { getMyActivities } from '@/features/my-page/my-activity/apis';
+import { getMyActivities } from '@/features/my-page/apis';
 import type { GetMyActivitiesQuery } from '@/features/my-page/types';
 import { QUERY_KEYS } from '@/shared/constants/queryKey';
 import { useUserStore } from '@/shared/stores/userStore';
